\name{allez-internal}

\alias{fn_isBinary}
\alias{fn_loadSetLibraries}
\alias{fn_loadPlatformLibraries}
\alias{fn_getDataEnv}
\alias{fn_getE.Globe}
\alias{fact}
\alias{sigma.fun}
\alias{sigma.fun.Nandi}
\alias{fn_getSetStats}
\alias{local.max}

\alias{ordMat}
\alias{allezplot}

\title{Internal allez functions}
\description{
  Internal allez functions; these are generally not to be called by the
  user.}

\author{Nandi Subhrangshu, Aimee Teo Broman}